+++ src/Neo.SmartContract.Testing/TestEngine.cs
@@ -530,8 +530,10 @@ namespace Neo.SmartContract.Testing
         /// Execute raw script
         /// </summary>
         /// <param name="script">Script</param>
+        /// <param name="initialPosition">Initial position (default=0)</param>
+        /// <param name="beforeExecute">Before execute</param>
         /// <returns>StackItem</returns>
-        public StackItem Execute(Script script)
+        public StackItem Execute(Script script, int initialPosition = 0, Action<ApplicationEngine>? beforeExecute = null)
         {
             // Store the script in current transaction
 
@@ -543,7 +545,7 @@ namespace Neo.SmartContract.Testing
 
             using var engine = new TestingApplicationEngine(this, Trigger, Transaction, snapshot, CurrentBlock);
 
-            engine.LoadScript(script);
+            engine.LoadScript(script, initialPosition: initialPosition);
 
             // Clean events, if we Execute inside and execute
             // becaus it's a mock, we can register twice
@@ -558,6 +560,7 @@ namespace Neo.SmartContract.Testing
 
             // Execute
 
+            beforeExecute?.Invoke(engine);
             var executionResult = engine.Execute();
 
             // Detach to static event
